apps/service_analyzer/requirements.txt
prometheus_client==0.24.1
python_json_logger==4.0.0
grpcio-tools==1.76.0
grpcio==1.76.0
aiohttp==3.13.3
aiokafka==0.13.0
dependency-injector==4.48.3
sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0
pydantic==2.12.5
pydantic_settings==2.12.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation-sqlalchemy
